Add Authentication and Authorization using JWT

+++ AirBnBApp/src/main/java/com/devshashi/AirBnBApp/service/HotelServiceImpl.java
@@ -5,13 +5,16 @@ import com.devshashi.AirBnBApp.dto.HotelInfoDTO;
 import com.devshashi.AirBnBApp.dto.RoomDTO;
 import com.devshashi.AirBnBApp.entity.Hotel;
 import com.devshashi.AirBnBApp.entity.Room;
+import com.devshashi.AirBnBApp.entity.User;
 import com.devshashi.AirBnBApp.exception.ResourceNotFoundException;
+import com.devshashi.AirBnBApp.exception.UnauthorizedException;
 import com.devshashi.AirBnBApp.repository.HotelRepository;
 import com.devshashi.AirBnBApp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -38,6 +41,10 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating a new hotel with name : {}", hotelDTO.getName());
         Hotel hotel = modelMapper.map(hotelDTO, Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         Hotel savedHotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID : {}", savedHotel.getId());
         return modelMapper.map(savedHotel, HotelDTO.class);
@@ -49,6 +56,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) throw new UnauthorizedException("This user does not own this hotel with id : " + id);
         return modelMapper.map(hotel, HotelDTO.class);
     }
 
@@ -58,6 +67,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) throw new UnauthorizedException("This user does not own this hotel with id : " + id);
         modelMapper.map(hotelDTO, hotel);
         hotel.setId(id);
         Hotel savedHotel = hotelRepository.save(hotel);
@@ -70,6 +81,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) throw new UnauthorizedException("This user does not own this hotel with id : " + id);
         for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -84,6 +97,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) throw new UnauthorizedException("This user does not own this hotel with id : " + id);
         hotel.setActive(true);
         for(Room room: hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
